Refactor contact pages to use client form components
